pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "SingleCellStochastics"
version = "0.1.0"
description = "An OU-Poisson model for metastatic cancer gene expression changes along a cell lineage tree."
authors = [
    {name = "Jiawei Xing", email = "xing@cshl.edu"},
    {name = "Stephen Staklinski", email = "staklins@cshl.edu"}
]
dependencies = [
    "biopython",
    "numpy",
    "scipy",
    "pandas",
    "statsmodels",
    "torch",
    "wandb",
    "joblib",
    "black",
    "matplotlib",
    "scanpy",
    "anndata",
    "scikit-learn"
]
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}

[project.urls]
"Homepage" = "https://github.com/Jiawei-Xing/SingleCellStochastics"

[project.scripts]
run-plasticity-test = "singlecellstochastics.plasticity:run_plasticity_test"
run-selection-test = "singlecellstochastics.selection:run_selection_test"
run-diff-test = "singlecellstochastics.ou_diff:run_diff_test"
run-reconst = "singlecellstochastics.reconstruct:run_reconst"
run-stochas-sim = "singlecellstochastics.stochas_sim:run_stochas_sim"
run-dea = "singlecellstochastics.dea:run_dea"
compare-results = "singlecellstochastics.results:compare_results"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
